Add ESlint and fix project structure.
