+++ services/matching-service/start.ts
@@ -1,10 +1,11 @@
 import express from "express";
-import { Server } from "socket.io";
 import { createServer } from "http";
+import { Server } from "socket.io";
 
-import { extractCategory, extractDifficulty, checkIdExists } from "./domain/match_validators";
-import { Match } from "./types";
-import { defineEventListeners } from "./events";
+// eslint-disable-next-line import/no-cycle
+import { defineEventListeners } from "./entry_points/api/events";
+
+import type { Match } from "./types";
 
 const app = express();
 const httpServer = createServer(app);
@@ -15,8 +16,7 @@ const port = process.env.PORT || 6001;
 defineEventListeners(io);
 
 httpServer.listen(port, () => {
-    console.log(`Matching service running on ${port}`)
+  console.log(`Matching service running on ${port}`);
 });
-
 
 export const queues: Map<string, Match> = new Map();
