+++ services/matching-service/types.ts
@@ -1,7 +1,6 @@
-
 export interface Match {
-    id: number;
-    difficulty: string;
-    category: string;
-    sockAddr: string;
+  id: number;
+  difficulty: string;
+  category: string;
+  sockAddr: string;
 }
